1.7.3

+++ Source/Baseline/Iti.Geolocation/GeoLocationHelper.cs
@@ -9,37 +9,37 @@ namespace Iti.Geolocation
         // Note: GPS 1's digit = 111 km
         //
         
-        public const decimal KilometersPerMile = 1.60934m;
-        public const decimal MilesPerKilometer = 1m / KilometersPerMile;
+        public const double KilometersPerMile = 1.60934;
+        public const double MilesPerKilometer = 1.0 / KilometersPerMile;
 
-        public static decimal MilesToKilometers(decimal miles)
+        public static double MilesToKilometers(double miles)
         {
             return miles * KilometersPerMile;
         }
 
-        public static decimal KilometersToMiles(decimal kilometers)
+        public static double KilometersToMiles(double kilometers)
         {
             return kilometers * MilesPerKilometer;
         }
 
-        public static void MinMaxRadiusMiles(GeoLocation geo, decimal radiusMiles, 
-            out decimal minLat, out decimal minLng, 
-            out decimal maxLat, out decimal maxLng)
+        public static void MinMaxRadiusMiles(GeoLocation geo, double radiusMiles, 
+            out double minLat, out double minLng, 
+            out double maxLat, out double maxLng)
         {
             var radiusKm = MilesToKilometers(radiusMiles);
             MinMaxRadiusKilometers(geo, radiusKm, out minLat, out minLng, out maxLat, out maxLng);
         }
 
-        public static void MinMaxRadiusKilometers(GeoLocation geo, decimal radiusKilometers,
-            out decimal minLat, out decimal minLng,
-            out decimal maxLat, out decimal maxLng)
+        public static void MinMaxRadiusKilometers(GeoLocation geo, double radiusKilometers,
+            out double minLat, out double minLng,
+            out double maxLat, out double maxLng)
         {
             minLat = minLng = maxLat = maxLng = 0;
 
             if (geo.Latitude == null || geo.Longitude == null)
                 return;
 
-            var latDiff = radiusKilometers / 111.0m;
+            var latDiff = radiusKilometers / 111.0;
 
             minLat = geo.Latitude.Value - latDiff;
             maxLat = geo.Latitude.Value + latDiff;
@@ -47,10 +47,10 @@ namespace Iti.Geolocation
             maxLng = geo.Longitude.Value + latDiff;
         }
 
-        private static double ToRadian(double val) { return val * (Math.PI / 180); }
+        private static double ToRadian(double val) { return val * (Math.PI / 180.0); }
         private static double DiffRadian(double val1, double val2) { return ToRadian(val2) - ToRadian(val1); }
 
-        public static decimal KilometerDistanceBetween(GeoLocation geo1, GeoLocation geo2)
+        public static double KilometerDistanceBetween(GeoLocation geo1, GeoLocation geo2)
         {
             const double earthRadiusInKilometers = 6367.0;
 
@@ -60,10 +60,10 @@ namespace Iti.Geolocation
             var lat2 = geo2.Latitude;
             var lng2 = geo2.Longitude;
 
-            return (decimal)(earthRadiusInKilometers * 2 * Math.Asin(Math.Min(1, Math.Sqrt((Math.Pow(Math.Sin((DiffRadian((double)lat1, (double)lat2)) / 2.0), 2.0) + Math.Cos(ToRadian((double)lat1)) * Math.Cos(ToRadian((double)lat2)) * Math.Pow(Math.Sin((DiffRadian((double)lng1, (double)lng2)) / 2.0), 2.0))))));
+            return (double)(earthRadiusInKilometers * 2 * Math.Asin(Math.Min(1, Math.Sqrt((Math.Pow(Math.Sin((DiffRadian((double)lat1, (double)lat2)) / 2.0), 2.0) + Math.Cos(ToRadian((double)lat1)) * Math.Cos(ToRadian((double)lat2)) * Math.Pow(Math.Sin((DiffRadian((double)lng1, (double)lng2)) / 2.0), 2.0))))));
         }
 
-        public static decimal MilesDistanceBetween(GeoLocation geo1, GeoLocation geo2)
+        public static double MilesDistanceBetween(GeoLocation geo1, GeoLocation geo2)
         {
             var kmDist = KilometerDistanceBetween(geo1, geo2);
             return KilometersToMiles(kmDist);
