+++ Source/Baseline/Iti.Geolocation/IGeolocator.cs
@@ -7,7 +7,7 @@ namespace Iti.Geolocation
     {
         GeoLocation Geocode(SimpleAddress simpleAddress);
         TimeZoneInfo TimezoneFor(GeoLocation location);
-        TimeZoneInfo TimezoneFor(decimal latitude, decimal longitude);
+        TimeZoneInfo TimezoneFor(double latitude, double longitude);
 
         double GetDrivingDistance(SimpleAddress from, SimpleAddress to);
     }
